feat(mobile): dé-duplication lot 2 — user_auth + team/tasks vers leopardo_core (progrès #5279)
- user_auth/data/user_auth_repository (unifié, formatage seul différait) + provider + 3 screens
- team_provider + task_provider (dépendances déjà dans le core)
- userAuthRepositoryProvider → core_providers commun ; google_sign_in + flutter_animate au core
- 14 instances dupliquées en moins (cumul 34 avec lot 1)
