+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 # Versioning : Semantic Versioning (semver.org) 
 
 ## [Unreleased]
+- **feat(mobile): dé-duplication lot 2 — grappe user_auth complète (repository + provider + 3 écrans) + team/tasks providers extraits vers leopardo_core (progrès #5279, suite #2601).** 7 paires supplémentaires éliminées (14 instances) : `features/user_auth/{data,providers,screens}` (repository unifié — les versions hr/manager ne différaient que par le formatage) + `team/providers/team_provider.dart` + `tasks/providers/task_provider.dart`. `userAuthRepositoryProvider` déplacé dans `leopardo_core/core/providers/core_providers.dart` (partie commune). Dépendances ajoutées au core : `google_sign_in ^7.2.0`, `flutter_animate ^4.5.2`. `flutter analyze` 0 issue ×3 packages ; tests 100/100 (hr 35, manager 34, employee 31). Cumul lots 1+2 : 17 paires éliminées (34 instances).
 - **feat(mobile): dé-duplication lot 1 — 10 écrans/providers partagés extraits vers leopardo_core (progrès #5279, suite #2601).** Providers `absence`, `cabinet`, `company_branding`, `evaluations`, `notification`, `payroll`, `salary_advance`, `schedule` + écrans `cabinet_screen`/`notification_list_screen` (100 % identiques entre `leopardo_hr`/`leopardo_manager` hors préfixe d'import) déplacés dans `leopardo_core/lib/features/**` ; nouveau `leopardo_core/core/providers/core_providers.dart` (partie commune — services + repositoryProviders des features du core) ; `core_providers.dart` des apps réduit à leurs providers spécifiques (contracts pour hr / ai_chat, vehicle_position, approvals pour manager) + re-export du core — tous les imports existants `package:<app>/core/providers/core_providers.dart` restent valides. Fix dépendabot : `flutter_secure_storage ^11` aligné dans les 2 apps (builds mobiles cassés sur main, core déjà en ^11). `image_picker ^1.2.3` ajouté au core (écran cabinet extrait). `flutter analyze` 0 issue ×3 packages, tests golden/smoke à jour. Réduction : 20 instances dupliquées éliminées (49 candidats → lots suivants : auth, attendance, team, tasks, user_auth, settings, onboarding).
 - **feat(marketing→accounting): lead qualifié → AccountingContact (Closes #5231).** `POST /marketing/leads/{lead}/qualify` (RBAC marketing/principal) : transition `status → qualified` + événement `MarketingLeadQualified` → listener `ConvertMarketingLeadToContact` → action `ConvertQualifiedLeadToContact` (crée `AccountingContact` pré-rempli source=marketing_lead, marketing_lead_id, NIF à compléter par le comptable) ; claim exclusif par tenant (`lockForUpdate`, 409 `LEAD_ALREADY_CLAIMED`, `converted_company_id` traçable). Lecture marketing read-only : `GET /marketing/leads/{lead}/contact` (404 cross-tenant). OpenAPI : 2 opérations + schéma `MarketingLeadContact` (redocly 0 erreur, couverture 747/747) ; SDK régénérés. 8 tests Feature `MarketingLeadConversionTest` (conversion, fallback nom email, exclusion cross-tenant, idempotence 409, RBAC marketing/principal vs employé/comptable, lecture, isolation).
 - **feat(accounting): CRUD contacts client/fournisseur + RBAC comptable (Closes #5222).** `AccountingContactController` (index/store/show/update/destroy) + FormRequests + routes `/api/v1/accounting/contacts` sous `api.manager:comptable,principal` ; isolation tenant par `BelongsToCompany` (404 cross-tenant, fail-closed #3727) ; NIF chiffré au repos (cast `encrypted`) ; filtres type/search + pagination bornée (per_page ≤ 100) ; payload filtré sur le fillable. OpenAPI : 5 opérations + schémas `AccountingContact`/`AccountingContactPayload` (751/751 routes couvertes, redocly 0 erreur) ; miroir `dev-hub/openapi/v1.yaml` + SDK JS/Python régénérés. 10 tests Feature `AccountingContactCrudTest` (CRUD, filtres, isolation tenant, RBAC comptable/principal/marketing/employé, validation 422, chiffrement NIF).
