+++ front/mobile_apps/leopardo_core/lib/features/tasks/providers/task_provider.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:leopardo_core/models/project_task.dart';
-import 'package:leopardo_hr/core/providers/core_providers.dart';
+import 'package:leopardo_core/core/providers/core_providers.dart';
 import 'package:leopardo_core/features/tasks/data/task_repository.dart';
 
 final taskRepositoryProvider = Provider<TaskRepository>((ref) {
