+++ front/mobile_apps/leopardo_core/lib/features/team/providers/team_provider.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:leopardo_hr/core/providers/core_providers.dart';
+import 'package:leopardo_core/core/providers/core_providers.dart';
 import 'package:leopardo_core/features/team/data/employee_repository.dart';
 import 'package:leopardo_core/models/employee.dart';
