+++ front/mobile_apps/leopardo_core/lib/features/user_auth/screens/user_login_screen.dart
@@ -8,7 +8,7 @@ import 'package:google_sign_in/google_sign_in.dart';
 import 'package:leopardo_core/core/theme/app_colors.dart';
 import 'package:leopardo_core/core/theme/app_typography.dart';
 import 'package:leopardo_core/core/widgets/demo_user_bottom_sheet.dart';
-import 'package:leopardo_manager/features/user_auth/providers/user_auth_provider.dart';
+import 'package:leopardo_core/features/user_auth/providers/user_auth_provider.dart';
 
 class UserLoginScreen extends ConsumerStatefulWidget {
   const UserLoginScreen({super.key});
