+++ front/mobile_apps/leopardo_core/pubspec.yaml
@@ -14,6 +14,8 @@ dependencies:
     sdk: flutter
   flutter_riverpod: ^3.4.2
   go_router: ^17.2.3
+  google_sign_in: ^7.2.0
+  flutter_animate: ^4.5.2
   dio: ^5.4.3+1
   flutter_secure_storage: ^11.0.0
   intl: ^0.20.2
