+++ front/mobile_apps/leopardo_hr/lib/app.dart
@@ -29,10 +29,10 @@ import 'package:leopardo_hr/features/payrolls/screens/payroll_list_screen.dart';
 import 'package:leopardo_core/features/cabinet/screens/cabinet_screen.dart';
 import 'package:leopardo_hr/features/settings/screens/settings_screen.dart';
 import 'package:leopardo_hr/features/team/screens/team_screen.dart';
-import 'package:leopardo_hr/features/user_auth/screens/user_register_screen.dart';
-import 'package:leopardo_hr/features/user_auth/screens/user_login_screen.dart';
+import 'package:leopardo_core/features/user_auth/screens/user_register_screen.dart';
+import 'package:leopardo_core/features/user_auth/screens/user_login_screen.dart';
 import 'package:leopardo_hr/features/user_auth/screens/user_home_screen.dart';
-import 'package:leopardo_hr/features/user_auth/screens/company_request_screen.dart';
+import 'package:leopardo_core/features/user_auth/screens/company_request_screen.dart';
 import 'package:leopardo_hr/features/contracts/screens/contract_screen.dart';
 import 'package:leopardo_hr/features/organigramme/screens/organigramme_screen.dart';
 import 'package:leopardo_hr/features/manager/screens/manager_attendance_monitoring_screen.dart';
