+++ front/mobile_apps/leopardo_hr/lib/features/tasks/screens/task_list_screen.dart
@@ -7,8 +7,8 @@ import 'package:leopardo_core/core/widgets/empty_state.dart';
 import 'package:leopardo_core/core/widgets/mobile_surface.dart';
 import 'package:leopardo_core/models/employee.dart';
 import 'package:leopardo_core/models/project_task.dart';
-import 'package:leopardo_hr/features/tasks/providers/task_provider.dart';
-import 'package:leopardo_hr/features/team/providers/team_provider.dart';
+import 'package:leopardo_core/features/tasks/providers/task_provider.dart';
+import 'package:leopardo_core/features/team/providers/team_provider.dart';
 import 'package:leopardo_core/core/i18n/device_locale.dart';
 
 class TaskListScreen extends ConsumerWidget {
