+++ front/mobile_apps/leopardo_hr/lib/features/team/screens/team_screen.dart
@@ -12,7 +12,7 @@ import 'package:leopardo_hr/features/auth/providers/auth_provider.dart';
 import 'package:leopardo_core/features/schedules/data/schedule_repository.dart';
 import 'package:leopardo_core/features/schedules/providers/schedule_provider.dart';
 import 'package:leopardo_core/features/team/data/employee_repository.dart';
-import 'package:leopardo_hr/features/team/providers/team_provider.dart';
+import 'package:leopardo_core/features/team/providers/team_provider.dart';
 import 'package:leopardo_core/models/employee.dart';
 import 'package:leopardo_core/core/utils/currency_format.dart';
 import 'package:leopardo_core/l10n/l10n.dart';
