+++ front/mobile_apps/leopardo_manager/lib/app.dart
@@ -29,10 +29,10 @@ import 'package:leopardo_core/features/cabinet/screens/cabinet_screen.dart';
 import 'package:leopardo_manager/features/settings/screens/settings_screen.dart';
 import 'package:leopardo_manager/features/team/screens/team_screen.dart';
 import 'package:leopardo_manager/features/tasks/screens/task_list_screen.dart';
-import 'package:leopardo_manager/features/user_auth/screens/user_register_screen.dart';
-import 'package:leopardo_manager/features/user_auth/screens/user_login_screen.dart';
+import 'package:leopardo_core/features/user_auth/screens/user_register_screen.dart';
+import 'package:leopardo_core/features/user_auth/screens/user_login_screen.dart';
 import 'package:leopardo_manager/features/user_auth/screens/user_home_screen.dart';
-import 'package:leopardo_manager/features/user_auth/screens/company_request_screen.dart';
+import 'package:leopardo_core/features/user_auth/screens/company_request_screen.dart';
 import 'package:leopardo_manager/features/ai_chat/screens/ai_chat_screen.dart';
 import 'package:leopardo_manager/features/vehicle_position/screens/vehicle_map_screen.dart';
 import 'package:leopardo_manager/features/approvals/screens/approval_screen.dart';
