+++ front/mobile_apps/leopardo_manager/lib/features/schedules/screens/schedule_list_screen.dart
@@ -9,7 +9,7 @@ import 'package:leopardo_core/models/employee.dart';
 import 'package:leopardo_manager/core/providers/core_providers.dart';
 import 'package:leopardo_core/features/schedules/data/schedule_repository.dart';
 import 'package:leopardo_core/features/schedules/providers/schedule_provider.dart';
-import 'package:leopardo_manager/features/team/providers/team_provider.dart';
+import 'package:leopardo_core/features/team/providers/team_provider.dart';
 import 'package:leopardo_core/core/widgets/mobile_list_glass_card.dart';
 
 class ScheduleListScreen extends ConsumerWidget {
