--- front/mobile_apps/leopardo_manager/lib/features/user_auth/providers/user_auth_provider.dart
@@ -1,106 +1,0 @@
-import 'package:flutter_riverpod/legacy.dart';
-import 'package:leopardo_core/core/api/api_exceptions.dart';
-import 'package:leopardo_manager/core/providers/core_providers.dart';
-import 'package:leopardo_manager/features/user_auth/data/user_auth_repository.dart';
-import 'package:leopardo_core/models/app_user.dart';
-
-class UserAuthState {
-  final bool isLoading;
-  final AppUser? user;
-  final String? error;
-
-  UserAuthState({this.isLoading = false, this.user, this.error});
-
-  UserAuthState copyWith({
-    bool? isLoading,
-    AppUser? user,
-    String? error,
-    bool clearError = false,
-    bool clearUser = false,
-  }) {
-    return UserAuthState(
-      isLoading: isLoading ?? this.isLoading,
-      user: clearUser ? null : (user ?? this.user),
-      error: clearError ? null : (error ?? this.error),
-    );
-  }
-}
-
-class UserAuthNotifier extends StateNotifier<UserAuthState> {
-  final UserAuthRepository _repository;
-
-  UserAuthNotifier(this._repository) : super(UserAuthState());
-
-  Future<void> checkAuth() async {
-    state = state.copyWith(isLoading: true);
-    final user = await _repository.checkAuth();
-    if (user != null) {
-      state = state.copyWith(isLoading: false, user: user);
-    } else {
-      state = state.copyWith(isLoading: false);
-    }
-  }
-
-  Future<bool> register({
-    required String firstName,
-    required String lastName,
-    required String email,
-    required String password,
-    String? phone,
-  }) async {
-    state = state.copyWith(isLoading: true, clearError: true);
-    try {
-      final data = await _repository.register(
-        firstName: firstName,
-        lastName: lastName,
-        email: email,
-        password: password,
-        phone: phone,
-      );
-      state = state.copyWith(isLoading: false, user: data['user'] as AppUser);
-      return true;
-    } catch (e) {
-      final msg =
-          e is ApiException ? e.message : 'Erreur lors de l\'inscription';
-      state = state.copyWith(isLoading: false, error: msg);
-      return false;
-    }
-  }
-
-  Future<bool> login(String email, String password) async {
-    state = state.copyWith(isLoading: true, clearError: true);
-    try {
-      final data = await _repository.login(email: email, password: password);
-      state = state.copyWith(isLoading: false, user: data['user'] as AppUser);
-      return true;
-    } catch (e) {
-      final msg = e is ApiException ? e.message : 'Identifiants invalides';
-      state = state.copyWith(isLoading: false, error: msg);
-      return false;
-    }
-  }
-
-  Future<bool> googleSignIn({required String idToken}) async {
-    state = state.copyWith(isLoading: true, clearError: true);
-    try {
-      final data = await _repository.googleSignIn(idToken: idToken);
-      state = state.copyWith(isLoading: false, user: data['user'] as AppUser);
-      return true;
-    } catch (e) {
-      final msg = e is ApiException ? e.message : 'Erreur connexion Google';
-      state = state.copyWith(isLoading: false, error: msg);
-      return false;
-    }
-  }
-
-  Future<void> logout() async {
-    await _repository.logout();
-    state = UserAuthState();
-  }
-}
-
-final userAuthProvider = StateNotifierProvider<UserAuthNotifier, UserAuthState>(
-  (ref) {
-    return UserAuthNotifier(ref.watch(userAuthRepositoryProvider));
-  },
-);
